+++ packages/plugins/google-antigravity/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aio-proxy/plugin-google-antigravity
 
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes