+++ packages/plugins/openai-chatgpt/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aio-proxy/plugin-openai-chatgpt
 
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.6.4
+  - @aio-proxy/types@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aio-proxy/plugin-sdk@0.6.3
+  - @aio-proxy/types@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes