+++ packages/server/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @aio-proxy/server
 
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies [[`08a579c`](https://github.com/aio-proxy/aio-proxy/commit/08a579cad9b5192820cd42f2cbb6ba18e0bc9e18)]:
+  - @aio-proxy/core@0.6.4
+  - @aio-proxy/i18n@0.6.4
+  - @aio-proxy/logger@0.6.4
+  - @aio-proxy/plugin-sdk@0.6.4
+  - @aio-proxy/types@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies [[`ba2aeae`](https://github.com/aio-proxy/aio-proxy/commit/ba2aeae4dfae3d932e2a22ac97d816b74d32a5ca)]:
+  - @aio-proxy/core@0.6.3
+  - @aio-proxy/i18n@0.6.3
+  - @aio-proxy/logger@0.6.3
+  - @aio-proxy/plugin-sdk@0.6.3
+  - @aio-proxy/types@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes